Time: 2 ms (96.68%), Space: 43.7 MB (94.01%) - LeetHub

+++ 0645-set-mismatch/0645-set-mismatch.java
@@ -2,12 +2,12 @@ class Solution {
     public int[] findErrorNums(int[] nums) {
         int[] res = new int[2];
         int sumDiff=0, n=nums.length;
-        long sumSqDiff=0;
+        int sumSqDiff=0;
         for(int i=0;i<n;i++){
             sumDiff += (nums[i] - (i+1));
             sumSqDiff += (nums[i]*nums[i] - (i+1)*(i+1));
         }
-        res[0] = (int)(sumSqDiff/sumDiff + sumDiff)/2;
+        res[0] = (sumSqDiff/sumDiff + sumDiff)/2;
         res[1] = res[0] - sumDiff;
         return res;
     }
